+++ projects/igniteui-angular/src/lib/drop-down/drop-down.component.spec.ts
@@ -4,8 +4,9 @@ import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { IgxToggleModule, IgxToggleDirective } from '../directives/toggle/toggle.directive';
 import { IgxDropDownItemComponent } from './drop-down-item.component';
-import { IgxDropDownComponent, IgxDropDownModule } from './drop-down.component';
+import { IgxDropDownComponent, IgxDropDownModule, ISelectionEventArgs } from './drop-down.component';
 import { IgxTabsComponent, IgxTabsModule } from '../tabs/tabs.component';
+import { UIInteractions } from '../test-utils/ui-interactions.spec';
 import { CancelableEventArgs } from '../core/utils';
 import { configureTestSuite } from '../test-utils/configure-suite';
 import { take } from 'rxjs/operators';
@@ -33,7 +34,8 @@ describe('IgxDropDown ', () => {
                 IgxDropDownInputTestComponent,
                 IgxDropDownImageTestComponent,
                 IgxDropDownTabsTestComponent,
-                DropDownWithValuesComponent
+                DropDownWithValuesComponent,
+                IgxDropDownSelectComponent
             ],
             imports: [
                 IgxDropDownModule,
@@ -53,7 +55,14 @@ describe('IgxDropDown ', () => {
             const button = fixture.debugElement.query(By.css('button')).nativeElement;
             const list = fixture.componentInstance.dropdown;
             let targetElement;
-            const mockObj = jasmine.createSpyObj('mockEvt', ['stopPropagation', 'preventDefault']);
+            const mockObj = {
+                key: '',
+                code: '',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
+            spyOn(mockObj, 'preventDefault');
+            spyOn(mockObj, 'stopPropagation');
             expect(list).toBeDefined();
             expect(list.items.length).toEqual(4);
 
@@ -67,7 +76,9 @@ describe('IgxDropDown ', () => {
             targetElement = fixture.debugElement.queryAll(By.css('.' + CSS_CLASS_DROP_DOWN_BASE))[0].parent;
             expect(targetElement).toBeDefined();
 
-            targetElement.triggerEventHandler('keydown.ArrowDown', mockObj);
+            mockObj.code = 'arrowdown';
+            mockObj.key = 'arrowdown';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
@@ -75,7 +86,9 @@ describe('IgxDropDown ', () => {
             expect(currentItem.componentInstance.index).toEqual(1);
             expect(list.selectedItem).toBeFalsy();
 
-            targetElement.triggerEventHandler('keydown.Space', mockObj);
+            mockObj.code = 'space';
+            mockObj.key = 'space';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             expect(list.selectedItem).toEqual(list.items[1]);
@@ -87,13 +100,17 @@ describe('IgxDropDown ', () => {
             targetElement = fixture.debugElement.queryAll(By.css('.' + CSS_CLASS_DROP_DOWN_BASE))[0].parent;
             currentItem = fixture.debugElement.queryAll(By.css('.' + CSS_CLASS_SELECTED))[0];
 
-            targetElement.triggerEventHandler('keydown.ArrowDown', mockObj);
+            mockObj.code = 'arrowdown';
+            mockObj.key = 'arrowdown';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             targetElement = fixture.debugElement.queryAll(By.css('.' + CSS_CLASS_DROP_DOWN_BASE))[0].parent;
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
 
-            targetElement.triggerEventHandler('keydown.Enter', mockObj);
+            mockObj.code = 'enter';
+            mockObj.key = 'enter';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             expect(list.collapsed).toEqual(true);
@@ -118,7 +135,14 @@ describe('IgxDropDown ', () => {
             const list = fixture.componentInstance.dropdownScroll;
             let targetElement;
             const listItems = list.items;
-            const mockObj = jasmine.createSpyObj('mockEvt', ['stopPropagation', 'preventDefault']);
+            const mockObj = {
+                key: '',
+                code: '',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
+            spyOn(mockObj, 'preventDefault');
+            spyOn(mockObj, 'stopPropagation');
             expect(list).toBeDefined();
             expect(list.items.length).toEqual(15);
 
@@ -129,14 +153,18 @@ describe('IgxDropDown ', () => {
             expect(currentItem.componentInstance.index).toEqual(0);
 
             targetElement = fixture.debugElement.query(By.css('.' + CSS_CLASS_DROP_DOWN_BASE)).parent;
-            targetElement.triggerEventHandler('keydown.ArrowDown', mockObj);
+            mockObj.code = 'arrowdown';
+            mockObj.key = 'arrowdown';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             expect(currentItem).toBeDefined();
             expect(currentItem.componentInstance.index).toEqual(1);
 
-            targetElement.triggerEventHandler('keydown.Space', mockObj);
+            mockObj.code = 'space';
+            mockObj.key = 'space';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             expect(list.selectedItem).toEqual(listItems[1]);
@@ -153,7 +181,9 @@ describe('IgxDropDown ', () => {
             expect(currentItem).toBeDefined();
             expect(currentItem.componentInstance.index).toEqual(4);
 
-            targetElement.triggerEventHandler('keydown.Escape', mockObj);
+            mockObj.code = 'Escape';
+            mockObj.key = 'Escape';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             expect(list.collapsed).toEqual(true);
@@ -169,34 +199,47 @@ describe('IgxDropDown ', () => {
             expect(list).toBeDefined();
             expect(list.items.length).toEqual(4);
 
-            const mockObj = jasmine.createSpyObj('mockEvt', ['stopPropagation', 'preventDefault']);
+            const mockObj = {
+                key: '',
+                code: '',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
             button.click();
             tick();
             fixture.detectChanges();
             let currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             targetElement = fixture.debugElement.query(By.css('.' + CSS_CLASS_DROP_DOWN_BASE)).parent;
-            targetElement.triggerEventHandler('keydown.ArrowDown', mockObj);
+            mockObj.code = 'ArrowDown';
+            mockObj.key = 'ArrowDown';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             expect(currentItem).toBeDefined();
             expect(currentItem.componentInstance.index).toEqual(1);
 
-            targetElement.triggerEventHandler('keydown.End', mockObj);
+            mockObj.code = 'End';
+            mockObj.key = 'End';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             expect(currentItem).toBeDefined();
             expect(currentItem.componentInstance.index).toEqual(3);
 
-            targetElement.triggerEventHandler('keydown.ArrowUp', mockObj);
+            mockObj.code = 'ArrowUp';
+            mockObj.key = 'ArrowUp';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             expect(currentItem).toBeDefined();
             expect(currentItem.componentInstance.index).toEqual(2);
 
-            targetElement.triggerEventHandler('keydown.Home', mockObj);
+            mockObj.code = 'Home';
+            mockObj.key = 'Home';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
             fixture.detectChanges();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
@@ -275,6 +318,84 @@ describe('IgxDropDown ', () => {
             expect(list.onSelection.emit).toHaveBeenCalledTimes(1);
             expect(list.onClosed.emit).toHaveBeenCalledTimes(1);
             expect(fixture.componentInstance.onSelection).toHaveBeenCalledTimes(1);
+        }));
+
+        it('Should check if selection event return the proper eventArgs', fakeAsync(() => {
+            const fixture = TestBed.createComponent(IgxDropDownTestComponent);
+            fixture.detectChanges();
+
+            const button = fixture.debugElement.query(By.css('button'));
+            const list = fixture.componentInstance.dropdown;
+            spyOn(list.onSelection, 'emit').and.callThrough();
+
+            UIInteractions.clickElement(button);
+            tick();
+            fixture.detectChanges();
+
+            UIInteractions.clickElement(list.items[3].element);
+            tick();
+            fixture.detectChanges();
+            const selectionArgs: ISelectionEventArgs = {
+                newSelection: list.items[3],
+                oldSelection: null,
+                cancel: false
+            };
+            expect(list.onSelection.emit).toHaveBeenCalledWith(selectionArgs);
+
+            UIInteractions.clickElement(button);
+            tick();
+            fixture.detectChanges();
+
+            UIInteractions.clickElement(list.items[1].element);
+            tick();
+            fixture.detectChanges();
+            expect(list.onSelection.emit).toHaveBeenCalledTimes(2);
+
+            tick();
+            fixture.detectChanges();
+            const selectionArgs1: ISelectionEventArgs = {
+                oldSelection: list.items[3],
+                newSelection: list.items[1],
+                cancel: false
+            };
+            expect(list.onSelection.emit).toHaveBeenCalledWith(selectionArgs1);
+        }));
+
+        it('Should check if selection event return the proper eventArgs if cancelled', fakeAsync(() => {
+            const fixture = TestBed.createComponent(IgxDropDownSelectComponent);
+            fixture.detectChanges();
+
+            const button = fixture.debugElement.query(By.css('button'));
+            const list = fixture.componentInstance.dropdown;
+            spyOn(list.onSelection, 'emit').and.callThrough();
+
+            UIInteractions.clickElement(button);
+            tick();
+            fixture.detectChanges();
+
+            UIInteractions.clickElement(list.items[3].element);
+            tick();
+            fixture.detectChanges();
+            const selectionArgs: ISelectionEventArgs = {
+                oldSelection: null,
+                newSelection: list.items[3],
+                cancel: true
+            };
+            expect(list.onSelection.emit).toHaveBeenCalledWith(selectionArgs);
+
+            UIInteractions.clickElement(button);
+            tick();
+            fixture.detectChanges();
+
+            UIInteractions.clickElement(list.items[1].element);
+            tick();
+            fixture.detectChanges();
+            const selectionArgs1: ISelectionEventArgs = {
+                oldSelection: null,
+                newSelection: list.items[1],
+                cancel: true
+            };
+            expect(list.onSelection.emit).toHaveBeenCalledWith(selectionArgs1);
         }));
 
         it('Should persist selection through scrolling', fakeAsync(() => {
@@ -400,7 +521,14 @@ describe('IgxDropDown ', () => {
             fixture.detectChanges();
             const button = fixture.debugElement.query(By.css('button')).nativeElement;
             const list = fixture.componentInstance.dropdown;
-            const mockObj = jasmine.createSpyObj('mockEvt', ['stopPropagation', 'preventDefault']);
+            const mockObj = {
+                key: 'arrowdown',
+                code: 'arrowdown',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
+            spyOn(mockObj, 'preventDefault');
+            spyOn(mockObj, 'stopPropagation');
             let targetElement;
             spyOn(list.onSelection, 'emit').and.callThrough();
             spyOn(list.onOpening, 'emit').and.callThrough();
@@ -417,14 +545,16 @@ describe('IgxDropDown ', () => {
             let currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             expect(currentItem).toBeDefined();
             targetElement = fixture.debugElement.query(By.css('.' + CSS_CLASS_DROP_DOWN_BASE)).parent;
-            targetElement.triggerEventHandler('keydown.ArrowDown', mockObj);
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             expect(currentItem).toBeDefined();
             expect(currentItem.componentInstance.index).toEqual(1);
-            targetElement.triggerEventHandler('keydown.Escape', mockObj);
+            mockObj.key = 'escape';
+            mockObj.code = 'escape';
+            targetElement.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
@@ -494,7 +624,12 @@ describe('IgxDropDown ', () => {
 
             let currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_SELECTED));
             expect(currentItem.componentInstance.index).toEqual(8);
-            currentItem.triggerEventHandler('keydown.Home', jasmine.createSpyObj('w/e', ['stopPropagation', 'preventDefault']));
+            currentItem.triggerEventHandler('keydown', {
+                key: 'Home',
+                code: 'Home',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            });
             tick();
 
             fixture.detectChanges();
@@ -509,7 +644,14 @@ describe('IgxDropDown ', () => {
             const button = fixture.debugElement.query(By.css('button')).nativeElement;
             const list = fixture.componentInstance.dropdownDisabled;
             const listItems = list.items;
-            const mockObj = jasmine.createSpyObj('mockEvt', ['stopPropagation', 'preventDefault']);
+            const mockObj = {
+                key: 'end',
+                code: 'end',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
+            spyOn(mockObj, 'stopPropagation');
+            spyOn(mockObj, 'preventDefault');
             expect(list).toBeDefined();
             expect(list.items.length).toEqual(13);
             button.click();
@@ -518,14 +660,16 @@ describe('IgxDropDown ', () => {
             fixture.detectChanges();
             let currentItem: DebugElement = fixture.debugElement.query(By.css('.' + CSS_CLASS_SELECTED));
             expect(list.items[8].isSelected).toBeTruthy();
-            currentItem.triggerEventHandler('keydown.End', mockObj);
+            currentItem.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
             expect(listItems[11].isFocused).toBeTruthy();
             currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_FOCUSED));
             expect(currentItem.componentInstance.index).toEqual(11);
-            currentItem.triggerEventHandler('keydown.ArrowDown', mockObj);
+            mockObj.key = 'ArrowDown';
+            mockObj.code = 'ArrowDown';
+            currentItem.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
@@ -540,7 +684,14 @@ describe('IgxDropDown ', () => {
             const button = fixture.debugElement.query(By.css('button')).nativeElement;
             const list = fixture.componentInstance.dropdownDisabled;
             const listItems = list.items;
-            const mockObj = jasmine.createSpyObj('mockEvt', ['stopPropagation', 'preventDefault']);
+            const mockObj = {
+                key: 'ArrowDown',
+                code: 'ArrowDown',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
+            spyOn(mockObj, 'stopPropagation');
+            spyOn(mockObj, 'preventDefault');
             expect(list).toBeDefined();
             expect(list.items.length).toEqual(13);
             button.click();
@@ -548,13 +699,15 @@ describe('IgxDropDown ', () => {
 
             fixture.detectChanges();
             let currentItem = fixture.debugElement.queryAll(By.css('.igx-drop-down__item'))[0];
-            currentItem.triggerEventHandler('keydown.ArrowDown', mockObj);
+            currentItem.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
             expect(list.items[10].isFocused).toBeTruthy();
             currentItem = fixture.debugElement.queryAll(By.css('.igx-drop-down__item'))[0];
-            currentItem.triggerEventHandler('keydown.ArrowUp', mockObj);
+            mockObj.key = 'ArrowUp';
+            mockObj.code = 'ArrowUp';
+            currentItem.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
@@ -619,7 +772,14 @@ describe('IgxDropDown ', () => {
             fixture.detectChanges();
             const button = fixture.debugElement.query(By.css('button')).nativeElement;
             const list = fixture.componentInstance.dropdownDisabled;
-            const mockObj = jasmine.createSpyObj('mockEvt', ['stopPropagation', 'preventDefault']);
+            const mockObj = {
+                key: '',
+                code: '',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
+            spyOn(mockObj, 'stopPropagation');
+            spyOn(mockObj, 'preventDefault');
             expect(list).toBeDefined();
             expect(list.items.length).toEqual(13);
             button.click();
@@ -628,7 +788,9 @@ describe('IgxDropDown ', () => {
             fixture.detectChanges();
             expect(list.items[8].isFocused).toEqual(true);
             const currentItem = fixture.debugElement.query(By.css('.' + CSS_CLASS_SELECTED));
-            currentItem.triggerEventHandler('keydown.ArrowDown', mockObj);
+            mockObj.code = 'ArrowDown';
+            mockObj.key = 'ArrowDown';
+            currentItem.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
@@ -639,7 +801,9 @@ describe('IgxDropDown ', () => {
 
             fixture.detectChanges();
             expect(list.items[10].isFocused).toEqual(true);
-            currentItem.triggerEventHandler('keydown.ArrowDown', mockObj);
+            mockObj.code = 'ArrowDown';
+            mockObj.key = 'ArrowDown';
+            currentItem.triggerEventHandler('keydown', mockObj);
             tick();
 
             fixture.detectChanges();
@@ -649,7 +813,13 @@ describe('IgxDropDown ', () => {
         it('Should select item and close on Enter keydown', fakeAsync(() => {
             const fixture = TestBed.createComponent(IgxDropDownWithScrollComponent);
             fixture.detectChanges();
-            const mockEvent = jasmine.createSpyObj('event', ['preventDefault']);
+            const mockObj = {
+                key: 'enter',
+                code: 'enter',
+                stopPropagation: () => {},
+                preventDefault: () => {}
+            };
+            spyOn(mockObj, 'preventDefault');
             const igxDropDown = fixture.componentInstance.dropdownScroll;
             igxDropDown.toggle();
             tick();
@@ -657,11 +827,11 @@ describe('IgxDropDown ', () => {
             expect(igxDropDown.selectedItem).toEqual(null);
             tick();
             const dropdownHandler = fixture.debugElement.query(By.css(CSS_CLASS_DROP_DOWN_BASE));
-            dropdownHandler.triggerEventHandler('keydown.Enter', mockEvent);
+            dropdownHandler.triggerEventHandler('keydown', mockObj);
             tick();
             expect(igxDropDown.collapsed).toEqual(true);
             expect(igxDropDown.selectedItem).toEqual(igxDropDown.items[0]);
-            expect(mockEvent.preventDefault).toHaveBeenCalledTimes(1);
+            expect(mockObj.preventDefault).toHaveBeenCalledTimes(1);
         }));
 
         it('should keep selection per instance', fakeAsync(() => {
@@ -699,8 +869,7 @@ describe('IgxDropDown ', () => {
             expect(dropdown.focusedItem).toEqual(null);
             inputElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
             tick();
-            expect(dropdown.selectItem).toHaveBeenCalledTimes(1);
-            expect(dropdown.selectItem).toHaveBeenCalledWith(null, jasmine.any(Object));
+            expect(dropdown.selectItem).toHaveBeenCalledTimes(0); // does not attempt to select item on keydown if DD is closed;
             expect(dropdown.selectedItem).toEqual(null);
             expect(dropdown.collapsed).toEqual(true);
             dropdown.toggle();
@@ -710,7 +879,7 @@ describe('IgxDropDown ', () => {
             const dropdownItem = dropdown.items[0];
             inputElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
             tick();
-            expect(dropdown.selectItem).toHaveBeenCalledTimes(2);
+            expect(dropdown.selectItem).toHaveBeenCalledTimes(1);
             expect(dropdown.selectItem).toHaveBeenCalledWith(dropdownItem, jasmine.any(Object));
             expect(dropdown.selectedItem).toEqual(dropdownItem);
             expect(dropdown.collapsed).toEqual(true);
@@ -1433,6 +1602,36 @@ class IgxDropDownTabsTestComponent {
     public onToggleClosed() { }
 }
 
+@Component({
+    template: `
+    <button (click)="toggleDropDown()">Toggle</button>
+    <igx-drop-down igxDropDownItemNavigation (onSelection)="onSelection($event)" [allowItemsFocus]="true"
+    [width]="'400px'" [height]="'400px'">
+        <igx-drop-down-item *ngFor="let item of items">
+            {{ item.field }}
+        </igx-drop-down-item>
+    </igx-drop-down>
+    `
+})
+class IgxDropDownSelectComponent {
+    @ViewChild(IgxDropDownComponent, { read: IgxDropDownComponent })
+    public dropdown: IgxDropDownComponent;
+
+    public items: any[] = [
+        { field: 'Nav1' },
+        { field: 'Nav2' },
+        { field: 'Nav3' },
+        { field: 'Nav4' }
+    ];
+
+    public toggleDropDown() {
+        this.dropdown.toggle();
+    }
+
+    public onSelection(eventArgs) {
+        eventArgs.cancel = true;
+    }
+}
 @Component({
     template: ` <input #inputElement [igxDropDownItemNavigation]="dropdownElement" class='test-input' type='text' value='Focus Me!'/>
     <igx-drop-down #dropdownElement [width]="'400px'" [height]="'400px'" [allowItemsFocus]="true">