+++ CHANGELOG.md
@@ -13,7 +13,8 @@ All notable changes for each version of this project will be documented in this 
     providers: [{ provide: DisplayDensityToken, useValue: { displayDensity: DisplayDensity.compact} }]
     })
     ```
-
+- `igx-drop-down`:
+    - Added a new boolean argument `cancel` to the `onSelection` `ISelectionEventArgs`. Its default value is false, in case it is set to true, the drop down selection is invalidated. 
 - `igxIcon`:
     - **Breaking change** `glyphName` property is removed from `IgxIconComponent`. For `Material` icons the icon name should be explicitly defined between the opening and closing tags. `Font Awesome` icons should use the `name` property now.
     - Added support for custom SVG icons. Register the SVG icons with the `IgxIconService` and use `IgxIconComponent`'s `name` and `fontSet` properties to visualize the icon.